ASoC: wcd9xxx-mbhc: Don't run consistency checks for headsets
Polling for the inserted plug interferes with the microphone pole, causing
the microphone on headsets to malfunction until the correct-plug worker
exits. This creates a ~2 second gap of time where users can experience
faulty microphone functionality (i.e. only loud static can he heard from
the microphone), even though the headset is properly detected and the
accompanying microphone is reported to Android.

Do not run the safety plug worker when headsets are detected in order to
fix the broken microphone functionality that appears for ~2 seconds. In
order to compensate for the lack of consistency checking, the plug type
is now polled a total of four times before the final result is actually
used to make a decision.

Signed-off-by: Sultanxda <sultanxda@gmail.com>