drivers: usb: dwc3: fix vendor edits
There are a few occurrences of

which isolate portions of code that should be compiled when building for bacon,
but they actually aren't, because that config entry isn't defined anywhere in the kernel tree.
So replace them in favor of CONFIG_MACH_FIND7OP, which is enabled in the defconfig.

Potentially dangerous, needs testing.